Add support of package_management in buildMyCube
* Building of JSON schema has been adapted to support nested arrays in objects.
* If previous configuration file existe (device-list) ask rather to
  delete it or keep it

+++ buildMyCube.sh
@@ -13,8 +13,8 @@ KALITE=False
 KALITE_LANG=""
 MEDIACENTER=False
 CSV_FILE_PATH=""
-ZIM=False
-ZIM_LIST=""
+PACKAGES_MANAGEMENT=False
+PACKAGES_LIST=""
 
 # configuration
 ANSIBLE_ETC="/etc/ansible/facts.d/"
@@ -113,7 +113,7 @@ function 3rd_party_app()
     CONFIGURE="own_config_file=True"
     KALITE=False
     MEDIACENTER=False
-    ZIM=False
+    PACKAGES_MANAGEMENT=False
 
     mkdir -p "$ANSIBLE_ETC"
 
@@ -150,7 +150,7 @@ function 3rd_party_app()
     fi
 
     if (dialog  --yesno "Do you want to download offline packages ?" 5 60) then
-        ZIM=True
+        PACKAGES_MANAGEMENT=True
 
         wget -O - http://catalog.ideascube.org/kiwix.yml > /tmp/kiwix.yml 2> /dev/null
         wget -O - http://catalog.ideascube.org/static-sites.yml >> /tmp/kiwix.yml 2> /dev/null
@@ -162,9 +162,7 @@ function 3rd_party_app()
                 --separate-output \
                 --ok-label "Add" \
                 --checklist "Select packages to install :" 100 100 40)
-        choices=$("${cmd[@]}" ${zim_files})
-
-        ZIM_LIST=`echo $choices | sed 's/ /","/g'`
+        declare -a PACKAGES_LIST=($("${cmd[@]}" ${zim_files}))
     fi
 
     echo -e "
@@ -178,17 +176,36 @@ function 3rd_party_app()
         \"idc_import\": {
             \"activated\": \"$MEDIACENTER\",
             \"content_name\": ["\"$CSV_FILE_PATH\""]
-        },
-        \"package_management\": {
-            \"activated\": \""$ZIM\"",
-            \"name\": [\""$ZIM_LIST"\"]
-        },
-        \"portal\": {
+        }," > /etc/ansible/facts.d/device_list.fact
+
+    if [ $PACKAGES_MANAGEMENT == "True" ]; then
+
+        echo -e "        \"package_management\":[{
+                    \"name\": \"${PACKAGES_LIST[0]}\",
+                    \"status\": \"present\"
+                " >> /etc/ansible/facts.d/device_list.fact
+
+        unset PACKAGES_LIST[0]
+        for PACKAGE in ${PACKAGES_LIST[@]}; do
+        echo -e "                },
+                {
+                    \"name\": \"$PACKAGE\",
+                    \"status\": \"present\"
+                " >> /etc/ansible/facts.d/device_list.fact
+        done
+
+        echo -e "                }
+        ]," >> /etc/ansible/facts.d/device_list.fact
+
+    fi
+
+    echo -e "        \"portal\": {
             \"activated\": \"True\"
-        }
+        },
+        \"kiwix_version\": \"1.0\"
     }
-}
-" > /etc/ansible/facts.d/device_list.fact
+}" >> /etc/ansible/facts.d/device_list.fact
+
         hostname $FULL_NAME
 }
 
@@ -261,6 +278,20 @@ internet_check
 
 [ $# -ne 0 ] || help
 
+if [[ -e /etc/ansible/facts.d/device_list.fact ]]; then
+    echo -n "[+] Local configuration file exist, would you like to delete it ? (y/n)" >&2
+    read response
+
+    case $response in
+        [OoYy]*)
+            rm -f /etc/ansible/facts.d/device_list.fact /tmp/device_list.fact
+        ;;
+        *)
+            cp /etc/ansible/facts.d/device_list.fact /tmp/device_list.fact
+        ;;
+    esac
+fi
+
 # Get argument from command line
 while [[ $# -gt 0 ]]
 do
@@ -315,7 +346,7 @@ do
             NAME="ideascube_project_name=$2"
             FULL_NAME=`echo "$2" | sed 's/_/-/g'`
 
-            wget https://raw.githubusercontent.com/ideascube/ansiblecube/$BRANCH/roles/set_custom_fact/files/device_list.fact -O /tmp/device_list.fact > /dev/null 2>&1
+            [[ -e /tmp/device_list.fact ]] || wget https://raw.githubusercontent.com/ideascube/ansiblecube/$BRANCH/roles/set_custom_fact/files/device_list.fact -O /tmp/device_list.fact > /dev/null 2>&1
 
             if [[ -z `grep -w "$FULL_NAME" /tmp/device_list.fact` ]] && [ "$TAGS" != "--tags master" ]
             then
